TTS: add channel adaptation layer for voice-bubble format conversion
Extract OGG/Opus transcoding into src/media/ensure-voice-format.ts so any
TTS provider (Edge, MiniMax, etc.) can produce voice bubbles on Telegram,
Feishu, and WhatsApp without needing native Opus output.

- Add src/media/ensure-voice-format.ts with ensureVoiceFormat() shared utility
- Refactor Discord ensureOggOpus() to delegate to the shared utility
- Update maybeApplyTtsToPayload() to transcode mp3->opus when the channel
  needs a voice bubble but the provider returned a non-opus file
- TTS providers remain format-agnostic; channel adaptation happens in the
  orchestration layer
